+++ package.json
@@ -17,7 +17,6 @@
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.14.8",
-    "@mui/lab": "^5.0.0-alpha.163",
     "@mui/material": "^5.14.8",
     "@mui/material-nextjs": "^5.15.5",
     "@mui/x-date-pickers": "^6.19.2",
