+++ src/app/page.tsx
@@ -1,7 +1,21 @@
-import MainPage from './panel/MainPage'
 import axios from 'axios'
+import PushAlertBanner from '@/app/panel/PushAlertBanner'
+import MainMobileView from '@/app/panel/main-page/MainMobileView'
+import Footer from '@/app/panel/main-page/Footer'
+import MainPcView from '@/app/panel/main-page/MainPcView'
+import PwaInstallBanner from '@/app/panel/PwaInstallBanner'
 
 export const dynamic = 'force-dynamic'
+
+export interface BeforeInstallPromptEvent extends Event {
+  readonly platforms: string[]
+  readonly userChoice: Promise<{
+    outcome: 'accepted' | 'dismissed'
+    platform: string
+  }>
+
+  prompt(): Promise<void>
+}
 
 export default async function Home({ searchParams }: { searchParams: any }) {
   let data
@@ -21,5 +35,20 @@ export default async function Home({ searchParams }: { searchParams: any }) {
     console.error('데이터를 불러오는데 실패했습니다.')
   }
 
-  return <MainPage initData={data} />
+  return (
+    <>
+      <PushAlertBanner />
+      {/* mobile view */}
+      <div className="mobile-layout">
+        <MainMobileView initData={data} />
+        <Footer />
+      </div>
+      {/* pc view */}
+      <div className="pc-layout">
+        <MainPcView initData={data} />
+        <Footer />
+      </div>
+      <PwaInstallBanner />
+    </>
+  )
 }
